Any reason not to use config.console_encoding ?
Using sys.std(out|err).encoding still causes problems when output is piped. Depending on how the next process handle the pipe, .encoding can be None.

If there's a particular reason for these to use sys.stderr.encoding, that'd be nice to use a config variable, and test for its non-"Noneness", as we currently do for config.console_encoding


git-svn-id: http://svn.wikimedia.org/svnroot/pywikipedia/branches/rewrite@6463 6a7f98fc-eeb0-4dc1-a6e2-c2c589a08aa6