Localisation updates from http://translatewiki.net.
Codes that fall back (for example sr to sr-ec) are now filled, and code code they fall back to are removed. There remains a sorting issue (causing fiu-vro to be sorted at vro), but this is much clearer than it was.
Rev 112439 in MediaWiki updated the code map for this (https://www.mediawiki.org/wiki/Special:Code/MediaWiki/112439).

git-svn-id: http://svn.wikimedia.org/svnroot/pywikipedia/branches/rewrite@9937 6a7f98fc-eeb0-4dc1-a6e2-c2c589a08aa6