Bugfix: actually login users that run login.py
Users were only logged in when they explicitly called -force, while
calling -force did not actually ask users that were logged in to
re-authenticate.

Calling login.py *without* -force now asks the user to log in (or displays
a message when this is already the case). Calling login.py *with* -force
displays a warning message with a description of how to logout (by deleting
the cookie file).



git-svn-id: http://svn.wikimedia.org/svnroot/pywikipedia/branches/rewrite@11342 6a7f98fc-eeb0-4dc1-a6e2-c2c589a08aa6